moved code to a new stdio command file
