--- signal-cli-script
@@ -1,1 +1,0 @@
-build/install/signal-cli/bin/signal-cli
